requirements.txt
requests~=2.32.3
python-dotenv~=1.1.0
Flask~=3.1.0
Flask-Cors~=5.0.1
PyYAML~=6.0.2
watchdog~=6.0.0
waitress~=3.0.2
